leader/requirements.txt
# AgSys Leader Dependencies
# Leader application for agricultural IoT system

# LoRa communication
spidev>=3.5
RPi.GPIO>=0.7.1

# Web API
flask>=2.3.0
flask-cors>=4.0.0

# Database
sqlite3  # Built-in

# Utilities
crcmod>=1.7
pyyaml>=6.0

# Logging
python-json-logger>=2.0.0
